# Conflicts:
#	src/main/webapp/homepage.jsp

+++ src/main/java/com/shop/actions/AdminAction.java
@@ -3,12 +3,15 @@ package com.shop.actions;
 import com.opensymphony.xwork2.ActionSupport;
 import com.shop.model.CustomOrder;
 import com.shop.model.Product;
+import com.shop.service.CustomOrderService;
 import com.shop.service.ProductService;
+import org.apache.struts2.interceptor.SessionAware;
 import org.springframework.beans.factory.annotation.Autowired;
 
 import java.util.ArrayList;
+import java.util.Map;
 
-public class AdminAction extends ActionSupport {
+public class AdminAction extends ActionSupport implements SessionAware {
     private ArrayList<Product> allProductList = new ArrayList<Product>();
     private ArrayList<CustomOrder> allOrdersList = new ArrayList<CustomOrder>();
     private String category;
@@ -19,13 +22,19 @@ public class AdminAction extends ActionSupport {
     private String description;
     private String image;
     private String errorMessage;
+    private Map<String, Object> session;
 
     @Autowired
     private ProductService productService;
 
-    //autowired pe customorderservice
+    @Autowired
+    private CustomOrderService customOrderService;
 
     public String execute() {
+        if (!session.get("loggedUser").equals("admin")) {
+            return "not_admin_user";
+        }
+
         if (id!=0 && name==null && description==null && totalQuantity==0 && price==0.0 && image==null && category==null){
             productService.deleteProduct(id);
         }
@@ -36,17 +45,17 @@ public class AdminAction extends ActionSupport {
             int addResult=productService.addProduct(name,description,totalQuantity,price,image,category);
             if (addResult==0){
                 errorMessage="Categoria nu exista!";
-                return "failed";
+                return "failure";
             }
             if (addResult==2){
                 errorMessage="Un produs cu acest nume exista deja!";
-                return "failed";
+                return "failure";
             }
         }
 
         allProductList = (ArrayList<Product>) productService.findAllProducts();
 
-        //allOrdersList = (ArrayList<CustomOrder>) customOrderService.findAllOrders();
+        allOrdersList = (ArrayList<CustomOrder>) customOrderService.findAllOrders();
 
         return "success";
     }
@@ -129,5 +138,10 @@ public class AdminAction extends ActionSupport {
 
     public void setAllOrdersList(ArrayList<CustomOrder> allOrdersList) {
         this.allOrdersList = allOrdersList;
+    }
+
+    @Override
+    public void setSession(Map<String, Object> session) {
+        this.session = session;
     }
 }
